ESDEV-3641 Remove isValidFieldName from Database class.
This method was used to "validate" an user provided value to use it later in a SQL query as the column name of the "order by" clause.
Field name validation is not done anymore as it was a very weak validation anyway and it would be too expensive to do a real , database platform independent validation. It would have to take into account reserved words, etc ...
Instead the column name part of "order by" clause is now quoted with DatabaseInterface::quoteIdentfier(), in BaseController::getSortingSql(), which is much more secure, as this string could always be safely (but maybe not successfully) used in a SQL query as column name in the "order by" clause.
(cherry picked from commit 949102e)